requirements.txt
appdirs==1.4.3
dateutil
Django==1.10.6
djangorestframework==3.5.4
packaging==16.8
pyparsing==2.2.0
six==1.10.0
